+++ Source/Events.Processing/Filters/AbstractFilterProcessor.cs
@@ -6,7 +6,7 @@ using Dolittle.DependencyInversion;
 using Dolittle.Logging;
 using Dolittle.Runtime.Events.Streams;
 
-namespace Dolittle.Runtime.Events.Processing
+namespace Dolittle.Runtime.Events.Processing.Filters
 {
     /// <summary>
     /// Represents an implementation of <see cref="IEventProcessor" />that processes the filtering of an event.
