+++ Source/Events.Processing/RuntimeServices.cs
@@ -4,6 +4,7 @@
 extern alias contracts;
 
 using System.Collections.Generic;
+using Dolittle.Runtime.Events.Processing.Filters;
 using Dolittle.Runtime.Services;
 using Dolittle.Services;
 using grpc = contracts::Dolittle.Runtime.Events.Processing;
